Sum of parameters using the variadic function

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -9,8 +9,8 @@
 int sum_them_all(const unsigned int n, ...)
 {
 
-	size_t sum = 0;
-	size_t i;
+	unsigned int sum = 0;
+	unsigned int i;
 
 	va_list holder;
 
@@ -22,10 +22,9 @@ int sum_them_all(const unsigned int n, ...)
 		{
 			return (0);
 		}
-
 		else
 		{
-			sum += va_arg(holder, const size_t);
+			sum += va_arg(holder, const unsigned int);
 		}
 	}
 	va_end(holder);
